+++ src/Microsoft.Kusto.ServiceLayer/DataSource/DataSourceIntellisense/ShowFunctionsResult.cs
@@ -1,0 +1,11 @@
+namespace Microsoft.Kusto.ServiceLayer.DataSource.DataSourceIntellisense
+{
+    public class ShowFunctionsResult
+    {
+        public string Name;
+        public string Parameters;
+        public string Body;
+        public string Folder;
+        public string DocString;
+    }
+}
